Move lmdb.Slab to separate file.
