+++ synapse/lib/cell.py
@@ -10,7 +10,7 @@ import synapse.common as s_common
 import synapse.telepath as s_telepath
 
 import synapse.lib.base as s_base
-import synapse.lib.lmdb as s_lmdb
+import synapse.lib.lmdbslab as s_lmdbslab
 import synapse.lib.const as s_const
 
 logger = logging.getLogger(__name__)
@@ -232,7 +232,7 @@ class Cell(s_base.Base, s_telepath.Aware):
 
         s_common.gendir(self.dirn, 'slabs')
         path = os.path.join(self.dirn, 'slabs', 'cell.lmdb')
-        self.slab = await s_lmdb.Slab.anit(path, map_size=s_const.gibibyte)
+        self.slab = await s_lmdbslab.Slab.anit(path, map_size=s_const.gibibyte)
         self.onfini(self.slab.fini)
 
     async def _initCellAuth(self):
